Added loading of external images back in. Cleaned up code substantially

+++ lib/image.php
@@ -8,39 +8,44 @@
 			'height' => 0,
 			'position' => 0,
 			'background' => 0,	
-			'file' => 0
+			'file' => 0,
+			'external' => false
 		);
 
 		## Mode 3: Resize Canvas
-		if(preg_match_all('/^3\/([0-9]+)\/([0-9]+)\/([1-9])\/([a-fA-f0-9]{3,6})\/(.+)$/i', $string, $matches, PREG_SET_ORDER)){
+		if(preg_match_all('/^3\/([0-9]+)\/([0-9]+)\/([1-9])\/([a-fA-f0-9]{3,6})\/(?:(0|1)\/)?(.+)$/i', $string, $matches, PREG_SET_ORDER)){
 			$param->mode = 3;
 			$param->width = $matches[0][1];
 			$param->height = $matches[0][2];			
 			$param->position = $matches[0][3];		
 			$param->background = $matches[0][4];
-			$param->file = $matches[0][5];
+			$param->external = (bool)$matches[0][5];			
+			$param->file = $matches[0][6];
 		}
 		
 		## Mode 2: Crop to fill
-		elseif(preg_match_all('/^2\/([0-9]+)\/([0-9]+)\/([1-9])(.+)$/i', $string, $matches, PREG_SET_ORDER)){
+		elseif(preg_match_all('/^2\/([0-9]+)\/([0-9]+)\/([1-9])\/(?:(0|1)\/)?(.+)$/i', $string, $matches, PREG_SET_ORDER)){
 			$param->mode = 2;
 			$param->width = $matches[0][1];
 			$param->height = $matches[0][2];			
-			$param->position = $matches[0][3];			
-			$param->file = $matches[0][4];
+			$param->position = $matches[0][3];
+			$param->external = (bool)$matches[0][4];			
+			$param->file = $matches[0][5];
 		}
 		
 		## Mode 1: Image resize
-		elseif(preg_match_all('/^1\/([0-9]+)\/([0-9]+)(.+)$/i', $string, $matches, PREG_SET_ORDER)){
+		elseif(preg_match_all('/^1\/([0-9]+)\/([0-9]+)\/(?:(0|1)\/)?(.+)$/i', $string, $matches, PREG_SET_ORDER)){
 			$param->mode = 1;
 			$param->width = $matches[0][1];
-			$param->height = $matches[0][2];			
-			$param->file = $matches[0][3];
+			$param->height = $matches[0][2];
+			$param->external = (bool)$matches[0][3];			
+			$param->file = $matches[0][4];
 		}
 		
 		## Mode 0: Direct displaying of image
-		elseif(preg_match_all('/^(.+)$/i', $string, $matches, PREG_SET_ORDER)){
-			$param->file = $matches[0][1];
+		elseif(preg_match_all('/^(?:(0|1)\/)?(.+)$/i', $string, $matches, PREG_SET_ORDER)){
+			$param->external = (bool)$matches[0][1];
+			$param->file = $matches[0][2];
 		}
 		
 		return $param;
@@ -61,16 +66,16 @@
 	define_safe('MODE_RESIZE_CROP', 2);
 	define_safe('MODE_CROP', 3);
 	
-	define_safe('CACHING', ($settings['image']['cache'] == 1 ? true : false));
+	define_safe('CACHING', ($param->external == false && $settings['image']['cache'] == 1 ? true : false));
 	
 	$meta = $cache_file = NULL;
 
-	$image_path = WORKSPACE . "/{$param->file}";
+	$image_path = ($param->external === true ? "http://{$param->file}" : WORKSPACE . "/{$param->file}");
 	
 	## Do cache checking stuff here
-	if(CACHING === true){
+	if($param->external !== true && CACHING === true){
 		
-	    $cache_file = sprintf('%s/%s_$s', CACHE, md5($_REQUEST['param'] . $quality), basename($image_path));	
+	    $cache_file = sprintf('%s/%s_%s', CACHE, md5($_REQUEST['param'] . $quality), basename($image_path));	
 
 		if(@is_file($cache_file) && (@filemtime($cache_file) < @filemtime($image_path))){ 
 			unlink($cache_file);
@@ -85,55 +90,40 @@
 	
 	####
 	
-	if($param->mode == MODE_NONE){
+	
+	if($param->external !== true && $param->mode == MODE_NONE){
 		
-		if(!$contents = @file_get_contents($image_path)){
+		if(!file_exists($image_path) || !is_readable($image_path)){
 			header('HTTP/1.0 404 Not Found');
 			trigger_error(__('Image <code>%s</code> could not be found.', array($image_path)), E_USER_ERROR);
 		}
 		
-		$meta = Image::meta($image_path);
-		Image::renderOutputHeaders($meta['type']);		
-		print $contents;
-		@imagedestroy($image);
+		$meta = Image::getMetaInformation($image_path);
+		Image::renderOutputHeaders($meta->type);
+		readfile($image_path);
 		exit();
-	} 
+	}
+	
 
 	try{
-		$image = Image::load($image_path, &$meta);
+		$method = 'load' . ($param->external === true ? 'External' : NULL);
+		$image = call_user_func_array(array('Image', $method), array($image_path));
 	}
 	catch(Exception $e){
 		header('HTTP/1.0 404 Not Found');
 		trigger_error($e->getMessage(), E_USER_ERROR);
 	}
 	
-
 	switch($param->mode){
 		
 		case MODE_RESIZE:
-			
-			include_once('filters/filter.resize.php');
-			
-			if($param->width == 0){
-				$r = (float)($param->height / $meta['height']);
-				$param->width = $meta['width'] * $r;
-			}
-			
-			elseif($param->height == 0){
-				$r = (float)($param->width / $meta['width']);
-				$param->height = $meta['height'] * $r;				
-			}
-		
-			//$result = $img->resize($param->width, $param->height, 'fill');
-			$image = FilterResize::run($image, $param->width, $param->height);
+			$image->applyFilter('resize', array($param->width, $param->height));
 			break;
 			
 		case MODE_RESIZE_CROP:
-
-			include_once('filters/filter.resize.php');
 		
-			$src_w = Image::width($image);
-			$src_h = Image::height($image);
+			$src_w = $image->Meta()->width;
+			$src_h = $image->Meta()->height;
 			
 			$dst_w = $param->width;
 			$dst_h = $param->height;
@@ -141,8 +131,8 @@
 			$src_r = ($src_w / $src_h);
 			$dst_r = ($dst_w / $dst_h);
 
-			if($src_r < $dst_r) $image = FilterResize::run($image, $dst_w, NULL);
-			else $image = FilterResize::run($image, NULL, $dst_h);			
+			if($src_r < $dst_r) $image->applyFilter('resize', array($dst_w, NULL));
+			else $image->applyFilter('resize', array(NULL, $dst_h));		
 			
 			/*
 				if($src_h < $param->height || $src_h > $param->height) ImageFilters::resize($image, NULL, $param->height);
@@ -151,16 +141,14 @@
 			*/
 
 		case MODE_CROP:
-			include_once('filters/filter.crop.php');
-			$image = FilterCrop::run($image, $param->width, $param->height, $param->position, $param->background);
+			$image->applyFilter('crop', array($param->width, $param->height, $param->position, $param->background));
 			break;
 	}
-
-	if(!Image::display($image, intval($settings['image']['quality']), true, $meta['type'])) trigger_error(__('Error generating image'), E_USER_ERROR);
+	
+	if(!$image->display(intval($settings['image']['quality']))) trigger_error(__('Error generating image'), E_USER_ERROR);
 	
 	if(CACHING && !is_file($cache_file)){ 
-		Image::save($image, $cache_file, intval($settings['image']['quality']), true, $meta['type']);
+		$image->save($cache_file, intval($settings['image']['quality']));
 	}
 	
-	@imagedestroy($image);
 	exit();
